Membuat Progress Screen v.1.1 (Memisahkan kode menjadi beberapa bagian)
